Split out supported domain functions
