+++ client/lib/gsuite/index.js
@@ -6,10 +6,8 @@ import { get, sortBy } from 'lodash';
 /**
  * Internal dependencies
  */
-import { isMappedDomainWithWpcomNameservers, isRegisteredDomain } from 'lib/domains';
 import { canDomainAddGSuite } from './can-domain-add-gsuite';
-import { hasGSuiteWithUs } from './has-gsuite-with-us';
-import { hasGSuiteWithAnotherProvider } from './has-gsuite-with-another-provider';
+import { getGSuiteSupportedDomains } from './gsuite-supported-domain';
 
 export { getAnnualPrice } from './get-annual-price';
 export { getMonthlyPrice } from './get-monthly-price';
@@ -25,6 +23,7 @@ export { getGSuiteMailboxCount } from './get-gsuite-mailbox-count';
 export { hasGSuiteWithUs } from './has-gsuite-with-us';
 export { hasGSuiteWithAnotherProvider } from './has-gsuite-with-another-provider';
 export { hasPendingGSuiteUsers } from './has-pending-gsuite-users';
+export { getGSuiteSupportedDomains, hasGSuiteSupportedDomain } from './gsuite-supported-domain';
 
 /**
  * Retrieves the first domain that is eligible to G Suite in this order:
@@ -49,37 +48,4 @@ export function getEligibleGSuiteDomain( selectedDomainName, domains ) {
 	);
 
 	return get( supportedDomains, '[0].name', '' );
-}
-
-/**
- * Filters a list of domains by the domains that eligible for G Suite.
- *
- * @param {Array} domains - list of domain objects
- * @returns {Array} - the list of domains that are eligible for G Suite
- */
-export function getGSuiteSupportedDomains( domains ) {
-	return domains.filter( function ( domain ) {
-		if ( hasGSuiteWithAnotherProvider( domain ) ) {
-			return false;
-		}
-
-		const isHostedOnWpcom =
-			isRegisteredDomain( domain ) && ( domain.hasWpcomNameservers || hasGSuiteWithUs( domain ) );
-
-		if ( ! isHostedOnWpcom && ! isMappedDomainWithWpcomNameservers( domain ) ) {
-			return false;
-		}
-
-		return canDomainAddGSuite( domain.name );
-	} );
-}
-
-/**
- * Given a list of domains does one of them support G Suite
- *
- * @param {Array} domains - list of domain objects
- * @returns {boolean} - Does list of domains contain a G Suited supported domain
- */
-export function hasGSuiteSupportedDomain( domains ) {
-	return getGSuiteSupportedDomains( domains ).length > 0;
 }
